refactor: indexInfo와 indexData 싱크를 service로 분리
